backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.13.3
pydantic==2.9.2
pydantic-settings==2.6.1
email-validator==2.1.0 
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
python-multipart==0.0.12
httpx==0.27.2
langchain==0.3.7
langgraph==0.2.50
google-cloud-aiplatform==1.73.0
langchain-google-vertexai==2.0.7
pypdf2==3.0.1